# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMANTIC - Change & Configuration API",
        "description": "API untuk CMDB, Change Management, Maintenance & Patch Management",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/v1/changes/{change}/decisions": {
            "post": {
                "tags": [
                    "Change Approval"
                ],
                "summary": "Memberikan keputusan approval untuk Change Request",
                "description": "POST /api/v1/changes/{change}/decisions",
                "operationId": "c3043bb6766f8abf348f4465bff3a537",
                "parameters": [
                    {
                        "name": "change",
                        "in": "path",
                        "description": "ID RFC / Change Request yang ingin di-approve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "stage",
                                    "decision"
                                ],
                                "properties": {
                                    "stage": {
                                        "description": "Level approval yang sedang memutuskan",
                                        "type": "string",
                                        "enum": [
                                            "kasi",
                                            "kabid",
                                            "kadis",
                                            "diskominfo"
                                        ],
                                        "example": "kabid"
                                    },
                                    "decision": {
                                        "description": "Keputusan approval",
                                        "type": "string",
                                        "enum": [
                                            "approve",
                                            "reject",
                                            "need_info"
                                        ],
                                        "example": "approve"
                                    },
                                    "note": {
                                        "description": "Catatan tambahan (opsional)",
                                        "type": "string",
                                        "example": "Setuju, pastikan backup full sebelum patch."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Keputusan approval berhasil direkam",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Decision recorded"
                                        },
                                        "approval": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "rfc_id": {
                                                    "type": "integer",
                                                    "example": 101
                                                },
                                                "level": {
                                                    "type": "string",
                                                    "example": "kabid"
                                                },
                                                "decision": {
                                                    "type": "string",
                                                    "example": "approve"
                                                },
                                                "reason": {
                                                    "type": "string",
                                                    "example": "Setuju, risiko sudah diterima."
                                                },
                                                "decided_at": {
                                                    "type": "string",
                                                    "example": "2025-11-15 10:30:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "404": {
                        "description": "RFC / Change Request tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/changes/{change}/approve": {
            "post": {
                "tags": [
                    "Change Approval"
                ],
                "summary": "Shortcut untuk menyetujui Change Request (decision=approve)",
                "description": "POST /api/v1/changes/{change}/approve",
                "operationId": "f720e1a4082b8cb1113a125df0355c0c",
                "parameters": [
                    {
                        "name": "change",
                        "in": "path",
                        "description": "ID RFC / Change Request yang ingin di-approve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "stage": {
                                        "type": "string",
                                        "enum": [
                                            "kasi",
                                            "kabid",
                                            "kadis",
                                            "diskominfo"
                                        ],
                                        "example": "kadis"
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Disetujui, window maintenance sesuai."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Keputusan approve tersimpan"
                    },
                    "404": {
                        "description": "RFC / Change Request tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/changes/{change}/reject": {
            "post": {
                "tags": [
                    "Change Approval"
                ],
                "summary": "Shortcut untuk menolak Change Request (decision=reject)",
                "description": "POST /api/v1/changes/{change}/reject",
                "operationId": "d58271673f5c5649427d537af7520d22",
                "parameters": [
                    {
                        "name": "change",
                        "in": "path",
                        "description": "ID RFC / Change Request yang ingin di-reject",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "stage": {
                                        "type": "string",
                                        "enum": [
                                            "kasi",
                                            "kabid",
                                            "kadis",
                                            "diskominfo"
                                        ],
                                        "example": "kabid"
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Ditolak, rencana rollback belum jelas."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Keputusan reject tersimpan"
                    },
                    "404": {
                        "description": "RFC / Change Request tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}